requirements.txt
Divisi2==2.2.5
Django==1.5.5
Django-Select2==4.2.2
South==0.8.2
Theano==0.6.0rc3
amqp==1.3.3
anyjson==0.3.3
argparse==1.2.1
beautifulsoup4==4.3.2
billiard==3.3.0.6
celery==3.1.3
csc-pysparse==1.1.1.4
csc-utils==0.6.7
dj-database-url==0.2.2
dj-static==0.0.5
django-attributes==0.2.0
django-crispy-forms==1.4.0
django-extensions==1.2.2
django-ratings==0.3.7
django-recommends==0.0.16
django-registration==1.0
django-tastypie==0.10.0
django-toolbelt==0.0.1
django-voting==0.2
gunicorn==18.0
kombu==3.0.4
mimeparse==0.1.3
networkx==1.8.1
numpy==1.7.1
psycopg2==2.5.1
pygraphviz==1.2
pymc==2.3
python-dateutil==2.1
pytz==2013.8
scikit-learn==0.14.1
scipy==0.12.0
six==1.4.1
static==0.4
wsgiref==0.1.2
